refactor: upgrade babel, migrate to jest
