+++ .travis.yml
@@ -4,4 +4,4 @@ node_js:
   - lts/*
 after_success:
   # - "npm run ci"
-  - "cat artifacts/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
+  - "cat coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
