+++ package.json
@@ -1,15 +1,14 @@
 {
   "name": "react-stickynode",
   "version": "3.0.2",
-  "description": "A performant and comprehensive React sticky",
+  "description": "A performant and comprehensive React sticky component",
   "main": "index.js",
   "scripts": {
     "build": "babel src --out-dir dist",
     "ci": "./tests/functional/saucelabs.sh",
     "lint": "eslint --cache --ext .js,.jsx . --fix",
     "prepublish": "npm run build",
-    "test": "jest",
-    "test:unit": "grunt unit",
+    "test": "jest --coverage tests/unit",
     "test:func": "grunt func"
   },
   "repository": {
@@ -43,6 +42,7 @@
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-react": "^7.9.4",
     "@babel/register": "^7.9.0",
+    "@testing-library/react": "^10.0.4",
     "babel-jest": "^26.0.1",
     "coveralls": "^3.1.0",
     "es5-shim": "^4.5.14",
@@ -56,23 +56,17 @@
     "grunt-contrib-connect": "^2.1.0",
     "grunt-contrib-watch": "^1.0.0",
     "grunt-saucelabs": "^9.0.1",
-    "grunt-shell": "^3.0.1",
     "grunt-webpack": "^3.1.3",
-    "istanbul": "^0.4.0",
     "jest": "^26.0.1",
     "jsdom": "^16.2.2",
-    "jsx-test": "^2.1.0",
     "minimist": "^1.2.5",
     "mocha": "^7.1.2",
-    "mockery": "^2.1.0",
     "pre-commit": "^1.0.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-test-renderer": "^16.13.1",
-    "sinon": "^9.0.2",
     "webpack": "^4.43.0",
-    "webpack-dev-server": "^3.10.3",
-    "xunit-file": "~1.0.0"
+    "webpack-dev-server": "^3.10.3"
   },
   "peerDependencies": {
     "react": "^0.14.2 || ^15.0.0 || ^16.0.0",
